# Improve zip file iteration performance