--- a/src/ICSharpCode.SharpZipLib/Core/InflaterPool.cs
+++ b/src/ICSharpCode.SharpZipLib/Core/InflaterPool.cs
@@ -5,35 +5,56 @@ using ICSharpCode.SharpZipLib.Zip.Compression;
 namespace ICSharpCode.SharpZipLib.Core
 {
 	/// <summary>
-	/// Pool Inflator instances as they can be costly due to byte array allocations.
+	/// Pool for <see cref="Inflater"/> instances as they can be costly due to byte array allocations.
 	/// </summary>
 	internal sealed class InflaterPool
 	{
-		public static InflaterPool Instance { get; } = new InflaterPool();
 		private readonly ConcurrentQueue<PooledInflater> noHeaderPool = new ConcurrentQueue<PooledInflater>();
 		private readonly ConcurrentQueue<PooledInflater> headerPool = new ConcurrentQueue<PooledInflater>();
+
+		internal static InflaterPool Instance { get; } = new InflaterPool();
 
 		private InflaterPool()
 		{
 		}
 
-		public Inflater Rent(bool noHeader = false)
+		internal Inflater Rent(bool noHeader = false)
 		{
+			if (SharpZipLibOptions.InflaterPoolSize <= 0)
+			{
+				return new Inflater(noHeader);
+			}
+
 			var pool = GetPool(noHeader);
-			var inf = pool.TryDequeue(out var inflater) ? inflater : new PooledInflater(noHeader);
-			inf.Reset();
+
+			PooledInflater inf;
+			if (pool.TryDequeue(out var inflater))
+			{
+				inf = inflater;
+				inf.Reset();
+			}
+			else
+			{
+				inf = new PooledInflater(noHeader);
+			}
+
 			return inf;
 		}
 
-		public void Return(Inflater inflater)
+		internal void Return(Inflater inflater)
 		{
+			if (SharpZipLibOptions.InflaterPoolSize <= 0)
+			{
+				return;
+			}
+
 			if (!(inflater is PooledInflater pooledInflater))
 			{
 				throw new ArgumentException("Returned inflater was not a pooled one");
 			}
 
 			var pool = GetPool(inflater.noHeader);
-			if (pool.Count < 10)
+			if (pool.Count < SharpZipLibOptions.InflaterPoolSize)
 			{
 				pooledInflater.Reset();
 				pool.Enqueue(pooledInflater);