--- a/src/ICSharpCode.SharpZipLib/Zip/Compression/PooledInflater.cs
+++ b/src/ICSharpCode.SharpZipLib/Zip/Compression/PooledInflater.cs
@@ -1,0 +1,14 @@
+﻿using ICSharpCode.SharpZipLib.Core;
+
+namespace ICSharpCode.SharpZipLib.Zip.Compression
+{
+	/// <summary>
+	/// A marker type for pooled version of an inflator that we can return back to <see cref="InflaterPool"/>.
+	/// </summary>
+	internal sealed class PooledInflater : Inflater
+	{
+		public PooledInflater(bool noHeader) : base(noHeader)
+		{
+		}
+	}
+}