@lahma 2023-08-12
It's always good to keep downstream consumers happy! 😉  I've force pushed a proposal of adding root level static class for global options, `SharpZipLibOptions`, which has now sole property of `InflaterPoolSize` which defaults to 0. Benchmark updated to set the value to 4 (allocations still down 70%).